Organise aws instrumentation code
